Add AbstractRecordService. Refactoring AbstractService and NewsGroupService.
