+++ src/main/java/su/svn/daybook/services/AbstractService.java
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2021.02.25 16:07 by Victor N. Skurikhin.
+ * This file was last modified at 2021.02.27 11:33 by Victor N. Skurikhin.
  * This is free and unencumbered software released into the public domain.
  * For more information, please refer to <http://unlicense.org>
  * AbstractService.java
@@ -12,34 +12,28 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.transaction.annotation.Transactional;
 import reactor.core.publisher.Mono;
-import su.svn.daybook.domain.dao.db.db.RecordDao;
 import su.svn.daybook.domain.model.DBUserOwnedEntry;
 import su.svn.daybook.domain.model.DBUuidEntry;
-import su.svn.daybook.domain.model.db.db.Record;
+
+import java.util.function.Function;
 
 @Slf4j
 public abstract class AbstractService<E extends DBUuidEntry> {
 
-    protected abstract RecordDao getRecordDao();
-
-    protected abstract Mono<Integer> insertEntry(E entry);
-
     @Transactional
-    public Mono<E> insert(Record record, final E entry) {
-        log.trace("insert({}, {})", record, entry);
+    public Mono<E> insert(Function<E, Mono<Integer>> daoInsertMethod, E entry) {
+        log.trace("insert({}, {})", daoInsertMethod, entry);
 
-        return getRecordDao().insert(record)
-                .doOnSuccess(i -> log.debug("insert {} count of record: {}", i, record))
-                .doOnError(e -> log.error("insertNewsEntry ", e))
-                .flatMap(i -> doOnInsert(i, entry))
-                .doOnSuccess(i -> log.debug("insert {} count of newsEntry: {}", i, entry))
-                .doOnError(e -> log.error("insertNewsEntry ", e));
+        return daoInsertMethod.apply(entry)
+                .doOnSuccess(i -> log.debug("insert {} count of entry: {}", i, entry))
+                .doOnError(e -> log.error("insert ", e))
+                .flatMap(i -> doOnInsert(i, entry));
     }
 
-    private Mono<E> doOnInsert(Integer i1, final E entry) {
-        return i1 != null && i1 == 1
-                ? insertEntry(entry).flatMap(i2 -> i2 == 1 ? Mono.just(entry) : Mono.empty())
-                : Mono.error(new RuntimeException(" doOnInsert catch i: " + i1));
+    private Mono<E> doOnInsert(Integer i, final E entry) {
+        return i != null && i == 1
+                ? Mono.just(entry)
+                : Mono.error(new RuntimeException(" doOnInsert catch i: " + i));
     }
 
     protected void setUserName(SecurityContext context, DBUserOwnedEntry entry) {
